third_party/WebKit/LayoutTests/animations/responsive/svg-responsive-to-timing-updates.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<svg>
  <stop id="target" offset="0"/>
</svg>
<script>
test(() => {
  var animation = target.animate({'svg-offset': '1'}, 1);
  assert_equals(target.offset.animVal, 0);
  animation.currentTime = 0.5;
  assert_equals(target.offset.animVal, 0.5);
  animation.currentTime = 0.75;
  assert_equals(target.offset.animVal, 0.75);
  animation.cancel();
  assert_equals(target.offset.animVal, 0);
}, 'SVG Web Animations should be responsive to changes in animation timing');
</script>
